Move shared author utilities into an authorutils module
check_orcid and latex2text now live in authorutils.py so that
authordb.py, the validated model layer, no longer imports from the
db2authors command line script. latex2text reuses a cached converter
instead of constructing LatexNodes2Text on every call, and repeated
conversions of the same string are memoized. The prototype tests
import bin modules at module scope, matching the PYTHONPATH=bin
convention used for pytest in CI.
