+++ bin/db2authors.py
@@ -29,6 +29,7 @@ from pathlib import Path
 from typing import Any, Self, TypeAlias
 
 import yaml
+from authorutils import check_orcid, latex2text
 
 try:
     from typing import Annotated
@@ -47,51 +48,6 @@ except ImportError:
     import dataclasses  # type: ignore[no-redef]
 
     OptStr: TypeAlias = str | None  # type: ignore[no-redef,misc]
-
-
-def latex2text(latex: str) -> str:
-    """Convert a LaTeX string into a plain text string.
-
-    Parameters
-    ----------
-    latex : `str`
-        Latex string to convert.
-
-    Returns
-    -------
-    plain : `str`
-        The plain text version.
-    """
-    from pylatexenc.latex2text import LatexNodes2Text
-
-    return LatexNodes2Text().latex_to_text(latex)
-
-
-def check_orcid(orcid: str | None) -> str | None:
-    """Validate and normalize an ORCID identifier.
-
-    Parameters
-    ----------
-    orcid : `str` or `None`
-        A bare dashed or 16-character ORCID identifier.
-
-    Returns
-    -------
-    normalized : `str` or `None`
-        The dashed identifier, or `None`.
-
-    Raises
-    ------
-    ValueError
-        Raised if the identifier is not in a recognized bare form.
-    """
-    if orcid is None:
-        return None
-    if re.fullmatch(r"\d{4}-\d{4}-\d{4}-\d{3}[0-9X]", orcid):
-        return orcid
-    if re.fullmatch(r"\d{15}[0-9X]", orcid):
-        return "-".join(orcid[i : i + 4] for i in range(0, len(orcid), 4))
-    raise ValueError(f"Given ORCiD does not match standard form: {orcid}")
 
 
 @dataclasses.dataclass(frozen=True)
